requirements.txt
Appium-Python-Client==4.5.0
